Move stabilize defs to package object
